refactor: update import paths for consistency and clarity across components and services
