+++ components/blog/BlogsList.tsx
@@ -17,7 +17,7 @@ import { Badge } from "@/components/ui/badge";
 import { Calendar, Clock, Languages } from "lucide-react";
 import type { Blog } from "@/types/blogs";
 import { BLOG_LANGUAGE_NAMES } from "@/types/blogs";
-import { LanguageBadge } from "./LanguageBadge";
+import { LanguageBadge } from "@/components/blog";
 import Image from "next/image";
 
 interface BlogCardProps {
