+++ components/visuals/ClientBackground.tsx
@@ -1,10 +1,10 @@
 /**
  * @author ColdByDefault
  * @copyright 2025 ColdByDefault. All Rights Reserved.
- */ 
+ */
 "use client";
 
-import { Background } from "./motion-background";
+import { Background } from "@/components/visuals";
 
 export default function ClientBackground() {
   return <Background />;
